Major overhaul: - Removed old prototype files - Added new Hero, Sidebar, and Typewriter components - Updated layout and main page - Removed old prototype files git push origin hero changed videos to mp4
