+++ app/page.tsx
@@ -1,9 +1,14 @@
 import Image from "next/image";
+import Sidebar from "./components/Sidebar";
+import Hero from "./components/Hero";
 
 export default function Home() {
   return (
     <main className="flex flex-col justify-center items-center w-full h-screen bg-cyan-950 px-4">
-      <h1 className="text-4xl md:text-6xl lg:text-8xl xl:text-9xl font-black text-white text-center">ViSiON</h1>
+      <Sidebar/>
+      <Hero/>
+      
+    
     </main>
   );
 }
