--- files/CONTRIBUTING.md
@@ -1,28 +1,0 @@
-# 🤝 Contributing to Vision
-
-Thanks for your interest in contributing!
-
-## 🚧 Workflow
-
-1. Fork the repository.
-2. Create a new branch:
-   - `feature/your-feature-name`
-   - `bugfix/your-bug-name`
-3. Write clean, well-commented code.
-4. Add or update tests if needed.
-5. Submit a Pull Request to the `develop` branch.
-6. Request a review from a teammate.
-
-## ✅ PR Checklist
-
-- [ ] Code builds without errors
-- [ ] Follows naming and branching conventions
-- [ ] Linting and formatting pass
-- [ ] Descriptive commit message and PR title
-
-## 🧪 Tools
-
-Use the following before submitting:
-```bash
-npm run lint
-npm run test
