--- files/SECURITY.md
@@ -1,25 +1,0 @@
-
----
-
-## 📄 `SECURITY.md`
-
-
-# 🔐 Security Policy
-
-## Reporting a Vulnerability
-
-If you find a security issue or vulnerability, please **do not open a public issue**.
-
-Instead, email us privately at: [vision.team@email.com](mailto:vision.team@email.com)
-
-## Security Best Practices
-
-- Never commit secrets (passwords, API keys, tokens).
-- Always use HTTPS for remote URLs.
-- Review dependencies for known vulnerabilities using tools like `npm audit`.
-
-## Disclosure Policy
-
-We respond to security reports within **48 hours**. Valid issues will be fixed ASAP with a proper changelog entry and patch release.
-
-Thanks for keeping our project secure!
