+++ public/files/CODE_OF_CONDUCT.md
@@ -1,0 +1,19 @@
+# 📜 Code of Conduct
+
+## Our Pledge
+
+We pledge to foster an open, welcoming, and respectful environment.
+
+## Our Standards
+
+- Use welcoming and inclusive language.
+- Be respectful of differing viewpoints and experiences.
+- Accept constructive criticism with grace.
+- Show empathy towards others.
+
+## Enforcement
+
+Violations can be reported confidentially to [vision.team@email.com](mailto:vision.team@email.com).  
+We commit to ensuring fair and prompt action in response.
+
+Thank you for helping create a positive environment!
