+++ public/files/prototype/about-us.html
@@ -1,0 +1,104 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Vision | About Us</title>
+  <link rel="stylesheet" href="css/About page styles.css"> 
+  <link rel="icon" type="image/png" href="images/Vision-logo.png" />
+</head>
+<body class="about-page">
+  <div class="about-page">
+    <div class="ripple-background" id="rippleBackground"></div>
+
+    <div class="main-content">
+      <nav>
+        <a href="index.html" >Home</a>
+        <a href="about-us.html">About</a>
+        <a href="projects.html">Projects</a>
+        <a href="Credit.html">Credit</a>
+        <a href="Contact-section01.html" >Contact</a>
+    </nav>
+      </nav>
+
+
+    <div class="about-founders-container">
+      <div class="background-element"></div>
+      <section class="about-founders-section">
+          <div class="image-wrapper">
+              <img src="images/pexels-pok-rie-33563-1726310.jpg" alt="About Us Image : Pexels" class="aquatic-about-image">
+          </div>
+          <div class="content-wrapper">
+              <p class="section-description">Vision is a dynamic group of junior developers navigating  the continuous learning curve of ever changing technology landscape . We believe that by building projects, experimenting with new technologies, and collaborating closely, we can gain the experience necessary to thrive in the dynamic tech industry. Our goal is not just to learn , but strive to bring new innovations to the tech world.</p>
+          </div>
+      </section>
+  </div>      
+
+  <!--/* Our Values */-->
+  <section>
+    <h2 style="color: white; text-align: center; font-size: 40px;"> Our Values </h2>
+
+    <div class="values-grid">
+        
+      <div class="values-box">
+        <div class="icon-container">
+            <svg class="icon" viewBox="0 0 24 24">
+                <path d="M18 11v2h4v-2h-4zm-2 6.61c.96.71 2.21 1.65 3.2 2.39.4-.53.8-1.07 1.2-1.6-.99-.74-2.24-1.68-3.2-2.4-.4.54-.8 1.08-1.2 1.61zM20.4 5.6c-.4-.53-.8-1.07-1.2-1.6-.99.74-2.24 1.68-3.2 2.4.4.53.8 1.07 1.2 1.6.96-.72 2.21-1.65 3.2-2.4zM4 9c-1.1 0-2 .9-2 2v2c0 1.1.9 2 2 2h1v4h2v-4h1l5 3V6L8 9H4zm11.5 3c0-1.33-.58-2.53-1.5-3.35v6.69c.92-.81 1.5-2.01 1.5-3.34z"></path>
+            </svg>
+        </div>
+        <h3 class="values-title">Innovation</h3>
+        <p class="values-description">We constantly seek new and better ways to [how you innovate].</p>
+    </div>
+
+  <div class="values-box blue">
+      <div class="icon-container">
+          <svg class="icon" viewBox="0 0 24 24">
+              <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 18c-4.41 0-8-3.59-8-8s3.59-8 8-8 8 3.59 8 8-3.59 8-8 8zm2-10c0 1.1-.9 2-2 2s-2-.9-2-2 .9-2 2-2 2 .9 2 2zm4 7.5c-.42-1.2-1.5-2.1-2.79-2.38C14.5 14.5 13.28 14 12 14s-2.5.5-3.21 1.12C7.5 15.4 6.42 16.3 6 17.5c-.53-1.35-.9-2.94-.9-4.5C5.1 8.04 8.14 5 12 5s6.9 3.04 6.9 6.98c0 1.48-.35 3.09-.9 4.52z"></path>
+          </svg>
+      </div>
+      <h3 class="values-title">Customer Focus</h3>
+      <p class="values-description">Our customers are at the heart of everything we do. We are committed to [how you prioritize customers].</p>
+  </div>
+
+  <div class="values-box">
+      <div class="icon-container">
+          <svg class="icon" viewBox="0 0 24 24">
+              <path d="M12 12c2.21 0 4-1.79 4-4s-1.79-4-4-4-4 1.79-4 4 1.79 4 4 4zm0 2c-2.67 0-8 1.34-8 4v2h16v-2c0-2.66-5.33-4-8-4z"></path>
+          </svg>
+      </div>
+      <h3 class="values-title">Integrity</h3>
+      <p class="values-description">We operate with honesty and transparency in all our interactions.</p>
+  </div>
+</div>
+  </section>
+  
+
+
+<footer>
+  <p class="footer-center-text">&copy; <span id="year"></span> Vision. All rights reserved.</p>
+</footer>
+
+
+
+
+  <script>
+    const rippleBackground = document.getElementById('rippleBackground');
+
+    function createRipple() {
+      const ripple = document.createElement('div');
+      ripple.classList.add('ripple');
+      const size = Math.random() * 100;
+      ripple.style.width = `${size}px`;
+      ripple.style.height = `${size}px`;
+      ripple.style.left = `${Math.random() * 100}vw`;
+      ripple.style.top = `${Math.random() * 100}vh`;
+      ripple.style.animationDelay = `${Math.random() * 5}s`;
+      rippleBackground.appendChild(ripple);
+      setTimeout(() => ripple.remove(), 6000);
+    }
+
+    setInterval(createRipple, 800);
+    document.getElementById("year").textContent = new Date().getFullYear();
+  </script>
+</body>
+</html>
