+++ public/files/prototype/css/projects.css
@@ -1,0 +1,238 @@
+/* Reset Styles */
+body {
+    margin: 0;
+    padding: 0;
+    display: flex;
+    min-height: 100vh;
+    background-color: #f4f4f4;
+    font-family: sans-serif;
+    color: #fff;
+}
+
+/* Ripple Background */
+.ripple-background {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: radial-gradient(circle, #748881, #4e7473);
+    overflow: hidden;
+    z-index: -1; /* Puts it behind content */
+}
+
+.ripple {
+    position: absolute;
+    width: 100px;
+    height: 100px;
+    border-radius: 50%;
+    background-color: rgba(62, 21, 21, 0.226);
+    animation: ripple-animation 6s infinite;
+    opacity: 0;
+}
+
+@keyframes ripple-animation {
+    0% {
+        transform: scale(0.5);
+        opacity: 0.8;
+    }
+    50% {
+        transform: scale(2);
+        opacity: 0.4;
+    }
+    100% {
+        transform: scale(4);
+        opacity: 0;
+    }
+}
+
+/* Sidebar Navigation */
+nav {
+    position: fixed;
+    left: 0;
+    top: 0;
+    height: 100vh;
+    background: rgba(12, 180, 192, 0);
+    width: 200px;
+    padding-top: 60px;
+    display: flex;
+    flex-direction: column;
+}
+
+nav a {
+    color: #eee;
+    text-decoration: none;
+    padding: 15px 20px;
+    font-weight: bold;
+    transition: background-color 0.3s ease;
+}
+
+nav a:hover {
+    background-color: rgba(0, 255, 204, 0.2);
+    color: #00ffcc;
+}
+
+.projects-header {
+    width: 100%;
+    text-align: center;
+    margin-bottom: 2rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.projects-header p {
+    color: #fff;
+    font-size: 1.2rem;
+    margin-top: -1rem; /* Adjust this to position the paragraph properly */
+}
+
+
+h2{
+    position: relative;
+    font-size: 8rem;
+    color: transparent;
+    -webkit-text-stroke: 0.1rem #00abf0;
+    background-image: linear-gradient(#00ffcc, #00abf0);
+    background-repeat: no-repeat;
+    -webkit-background-clip: text;
+    background-position: -66rem 0;
+    /* set the minimum position value so that the color did not show up. if you add padding then add extra padding value also for example here at first it was -64rem later I add 1rem padding both left and right side of the text. then we have to change the -66rem position */
+    padding: 0 1rem;
+    animation: animate 6s linear infinite;
+}
+
+h2::before{
+    content: '';
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 0%;
+    height: 100%;
+    border-right: 4px solid #00abf0;
+    animation: lineAnimate 6s linear infinite;
+}
+
+@keyframes animate {
+    0%, 10%, 100%{
+        background-position: -66rem 0;
+    }
+    80%,90%{
+        background-position: 0 0;
+    }
+}
+
+@keyframes lineAnimate {
+    0%, 10%, 100%{
+        width: 0%;
+    }
+    80%,90%{
+        width: 100%;
+    }
+}
+
+/* Content Wrapper */
+.content-wrapper {
+    margin-left: 20px; /* To avoid overlapping with sidebar */
+    flex-grow: 1;
+    padding: 50px;
+    position: relative;
+    z-index: 2;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+::content{
+    align-items: center;
+}
+
+/* Grid Layout */
+.about-grid-3col {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+    gap: 20px;
+    padding: 20px;
+    max-width: 100px;
+    width: 100%;
+}
+
+.about-section {
+    text-align: left;
+    padding: 20px;
+    border: 1px solid #eee;
+    border-radius: 5px;
+    color: #fff;
+}
+
+.about-section h2 {
+    margin-top: 0;
+    color: #fff;
+    margin-bottom: 15px;
+    text-align: center;
+}
+
+/* Responsive Grid */
+@media (max-width: 768px) {
+    .about-grid-3col {
+        grid-template-columns: 1fr;
+    }
+}
+
+.container{
+    width: 100%;
+    height: 35vh;
+    margin-top: 1,2rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 1rem;
+    /* add 3d representation to the container div element */
+    transform-style: preserve-3d;
+    transform: perspective(1000px);
+}
+
+.container .item{
+    width: 10rem;
+    height: 20rem;
+    /* add brightness animation effect */
+    transition: 0.5s;
+    filter: brightness(0);
+    cursor: pointer;
+}
+
+.container .item:hover{
+    filter: brightness(1);
+    transform: translateZ(200px);
+}
+
+/* add hover animation to the right item */
+.container .item:hover + *{
+    filter: brightness(0.6);
+    transform: translateZ(150px) rotateY(40deg);
+}
+
+.container .item:hover + * + *{
+    filter: brightness(0.4);
+    transform: translateZ(70px) rotateY(20deg);
+}
+
+/* for left of the hover */
+/* +* represent the right element but for left we don't have any thing like this */
+
+.container .item:has(+ *:hover){
+    filter: brightness(0.6);
+    transform: translateZ(150px) rotateY(-40deg);
+}
+
+.container .item:has(+ * + *:hover){
+    filter: brightness(0.4);
+    transform: translateZ(70px) rotateY(-20deg);
+}
+
+.item img{
+    width: 100%;
+    height: 100%;
+    object-fit: relatively cover;
+}
