+++ public/files/prototype/css/styles.css
@@ -1,0 +1,126 @@
+{
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+body {
+    font-family: 'Arial', sans-serif;
+    color: rgb(52, 49, 49);
+    overflow: hidden;
+    display: flex;/* Enable flexbox for layout */
+    background-attachment: fixed;
+    background-size: cover;
+    transition: background 0.5s ease;
+}
+
+.icon-style{
+    width:150px;
+    height: 150px;
+    
+    
+}
+
+#video-background {
+    position: fixed;
+    right: 0;
+    bottom: 0;
+    min-width: 100%;
+    min-height: 100%;
+    z-index: -1;
+}
+
+nav {
+    
+    width: 200px; /* Adjust sidebar width as needed */
+    height: 100vh; /* Full height sidebar */
+    padding-top: 180px; /* Add some top padding for spacing */
+    background: rgba(0, 0, 0, 0); /* Semi-transparent background */
+    z-index: 10; /* Ensure it's above the video and content */
+    display: flex;
+    flex-direction: column; /* Stack links vertically */
+    
+} 
+
+nav a {
+    color: #eee;
+    text-decoration: none;
+    padding: 15px 20px;
+    width: 100%;
+    box-sizing: border-box;
+    font-weight: bold;
+    transition:  0.3s ease;
+}
+
+nav a:hover {
+    background-color: rgba(0, 255, 204, 0.2);
+    color: #00ffcc;
+}
+
+.content {
+    margin-left: 0; /* Remove sidebar space when closed */
+    transition: margin-left 0.3s ease
+}
+
+nav[style*="left: 0"] ~ .content {
+    margin-left: 200px;
+}
+h1 {
+    font-size: 3rem;
+    margin-bottom: 1rem;
+    font-weight: 300;
+}
+
+
+
+p {
+    font-size: 1.2rem;
+    margin-bottom: 2rem;
+    max-width: 600px;
+    line-height: 1.6;
+}
+
+.btn {
+    padding: 12px 30px;
+    background: transparent;
+    
+   
+    text-decoration: none;
+    font-size: 1rem;
+    transition: all 0.3s ease;
+}
+
+.btn:hover {
+    background: rgb(224, 221, 221);
+    color: rgb(248, 247, 247);
+}
+
+/* Add to the end of styles.css */
+.page-nav-buttons button {
+    padding: 10px 15px;
+    background: rgba(0, 255, 204, 0.2);
+    border: 1px solid #00ffcc;
+    border-radius: 5px;
+    color: #00ffcc;
+    cursor: pointer;
+    transition: all 0.3s ease;
+}
+
+.page-nav-buttons button:hover {
+    background: rgba(0, 255, 204, 0.4);
+}
+
+@media (max-width: 768px) {
+    .page-nav-buttons {
+        bottom: 10px;
+        right: 10px;
+    }
+    
+    .page-nav-buttons button {
+        padding: 8px 12px;
+    }
+}
+
+html {
+    scroll-behavior: smooth;
+}
