+++ public/files/prototype/index.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Vision 2025</title>
+    <link rel="stylesheet" href="css/styles.css">
+</head>
+<body>
+    <video autoplay muted loop id="video-background">
+        <source src="vid/bga.mp4" type="video/mp4">
+        Your browser does not support HTML5 video.
+    </video>
+    
+    <header>
+        
+
+        <nav id="sidebar" >
+            <a href="index.html" class="smooth-link">Home</a>
+            <a href="about-us.html" class="smooth-link">About</a>
+            <a href="projects.html" class="smooth-link">Projects</a>
+            <a href="Credit.html" class="smooth-link">Credit</a>
+            <a href="Contact-section01.html" >Contact</a>
+        </nav>
+        <div class="content-wrapper">
+           
+        </div>
+
+    </header>
+    
+    <main class="content">
+        
+    </main>
+          
+       
+    
+    <script src="js/script.js"></script>
+</body>
+</html>
