+++ public/files/prototype/projects.html
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="css/projects.css">
+    <title>Projects</title>
+</head>
+<body>
+    <!-- Ripple Background -->
+    <div class="ripple-background"></div>
+
+    <!-- Sidebar Navigation -->
+    <nav id="sidebar">
+        <a href="index.html" class="smooth-link">Home</a>
+        <a href="about-us.html" class="smooth-link">About</a>
+        <a href="projects.html" class="smooth-link">Projects</a>
+        <a href="Credit.html" class="smooth-link">Credit</a>
+        <a href="Contact-section01.html" class="smooth-link">Contact</a>
+    </nav>
+
+    <!-- Main Content -->
+    <div class="content-wrapper">
+        <h2>Projects</h2>
+        <p>Explore our latest projects and initiatives.</p>
+        
+    
+    <div class="container">
+        <div class="item"><img src="img/2.jpeg" alt=""></div>
+        <div class="item"><img src="img/3.jpeg" alt=""></div>
+        <div class="item"><img src="img/4.jpeg" alt=""></div>
+        <div class="item"><img src="img/5jpeg.jpeg" alt=""></div>
+        <div class="item"><img src="img/6.jpeg" alt=""></div>
+        <div class="item"><img src="img/7.jpeg" alt=""></div>
+    
+        <div class="item"><img src="https://i.pinimg.com/474x/fe/ef/92/feef926836fbc7130733bf83f74ac26f.jpg" alt=""></div>
+        <div class="item"><img src="https://i.pinimg.com/236x/e5/69/a5/e569a52cc49ee81595eba5e077a09517.jpg" alt=""></div> 
+    </div>
+  
+
+    <script src="js/projects.js"></script>
+</body>
+</html>
